requirements.txt
# Python核心依赖
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 数据获取
tushare>=1.2.89
akshare>=1.8.0
yfinance>=0.2.0

# 数据处理
scikit-learn>=1.0.0
ta-lib>=0.4.0  # 需要先安装TA-Lib C库

# 深度学习
tensorflow>=2.8.0
# torch>=1.10.0  # 如果使用PyTorch
# transformers>=4.20.0  # 如果使用Hugging Face模型

# 机器学习
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Web框架
fastapi>=0.85.0
uvicorn>=0.18.0
flask>=2.2.0
flask-cors>=3.0.10

# 数据库
sqlalchemy>=1.4.0
pymysql>=1.0.0
psycopg2-binary>=2.9.0
redis>=4.3.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
mplfinance>=0.12.0

# 配置管理
pyyaml>=6.0
python-dotenv>=0.20.0

# 工具库
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
openpyxl>=3.0.0

# 日期时间
python-dateutil>=2.8.0
pytz>=2022.1

# 日志
loguru>=0.6.0

# 任务调度
apscheduler>=3.9.0

# 数据验证
pydantic>=1.10.0

# 进度条
tqdm>=4.64.0

# Jupyter支持
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.7.0

# 测试
pytest>=7.1.0
pytest-cov>=3.0.0

# 代码质量
black>=22.6.0
flake8>=5.0.0
pylint>=2.14.0
